Replaced "seL4-project" remote with "sel4proj"
Change-Id: I899db75930eb6a917935fc25c44bf590d8338cf7

sel4-project was the original project name when we used github
internally and externally. When we changed to using bitbucket
internally, we had to change our internal project name to sel4proj,
as bitbucket doesn't allow hyphens in repo URIs. We've since changed
the name of the project on our external github account to sel4proj
as well, to prevent the need to deal with different project names
internally and externally.